==> R/01-2_OISST_land_seaice_mask.Rmd <==
---
title: "To mask seaice and land from OISST dataset"
author: "cywhale"
date: "`r Sys.Date()`"
output:
  html_document:
    theme: flatly
    highlight: tango
    self-contained: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(tidy = FALSE)
options(htmltools.dir.version = FALSE)
options(useFancyQuotes = FALSE)
```

R code modified from https://github.com/mjacox/Thermal_Displacement/blob/master/make_oisst_masks.m

Seaice mask from OISST data (R/01-1_OISST_monthly.Rmd)

  - Use monthly seaice mask data, must run R/01-1_OISST_monthly.Rmd first, with larger year range, e.g yrng=1982:2019


Data source: https://ldas.gsfc.nasa.gov/sites/default/files/ldas/gldas/VEG/GLDASp5_landmask_025d.nc4 (0.25 degree, 1400x600 -59.875 - 89.875)

  - Data from NASA Goddard Earth Sciences Data and Information Services Center (GES DISC)
  
  - Metadata: https://ldas.gsfc.nasa.gov/gldas/vegetation-class-mask

``` {r landmask_from_gldas, warning=FALSE, message=FALSE, results='hide'}
if (!require("curl")) install.packages("curl"); library(curl)
if (!require("magrittr")) install.packages("magrittr"); library(magrittr)
if (!require("data.table")) install.packages("data.table"); library(data.table)
if (!require("stars")) install.packages("stars"); library(stars)
if (!require("abind")) install.packages("abind"); library(abind)
if (!require("dplyr")) install.packages("dplyr"); library(dplyr)
if (!require("ggplot2")) install.packages("ggplot2"); library(ggplot2)
if (!require("ggExtra")) install.packages("ggExtra"); library(ggExtra)
if (!require("viridis")) install.packages("viridis"); library(viridis)
if (!require("RColorBrewer")) install.packages("RColorBrewer"); library(RColorBrewer)

landmask_file = "../data_src/mapfiles/GLDASp4_landmask_025d.nc4"
if (!file.exists(landmask_file)) {
  tryCatch({
    curl::curl_download("https://ldas.gsfc.nasa.gov/sites/default/files/ldas/gldas/VEG/GLDASp5_landmask_025d.nc4", destfile = landmask_file)
  }, error = function(e) paste0(e))
} 

lsm <- read_stars(landmask_file) #0 = water; 1 = land #downloaded from 
dt0 <- as.data.table(lsm)
dt0[, time:=NULL] 
setnames(dt0, 3, "landmask")
dt0[, seamask:= as.integer(!landmask)] #Note: seamask 0 is land

ggplot() +  
  geom_tile(data=dt0, aes(x=x, y=y, fill=factor(seamask))) + 
  scale_fill_viridis_d(direction=-1) +
  coord_sf() + #xlim=c(-180, 180), ylim=c(-90, 90)) +
  scale_x_continuous(limits = c(-180, 180), expand = c(0, 0)) + 
  scale_y_continuous(limits = c(range(dt0$y)[1], range(dt0$y)[2]), expand = c(0, 0)) +
  xlab(NULL) + ylab(NULL)
```
To evaluate thermal displacement under marine heatwaves, in Jacox et al. (2020):

  - Thermal displacements were constrained so that unrealistic paths through land barriers (for example, entering or exiting inland seas, crossing continents between ocean basins) did not alter the large-scale patterns presented here. However, paths that interacted with land were allowed if they represented realistic displacements (for example, along the California coast in Fig. 3a); in such cases, the reported thermal displacements underestimate the true distance travelled by an oceanic pathway. Regions for which a displaced thermal habitat is sometimes unreachable include inland seas as well as gulfs, bays, and seas that are bounded by land masses on the poleward side.
  
   - https://www.nature.com/articles/s41586-020-2534-z#Sec2


That's why here we need to mask land, seaice, and some inland seas.


```{r use_monthly_sea_icemask, warning=FALSE, message=FALSE, results='hide'}
## Note: Time-consuming!! So if you don't want to re-calculate, just use saved data and re-load it 
Recalculate <- FALSE
if (!Recalculate) {
  #load("../data_src/stats/sea_icemask_1982_2019avg_nc.RData") #old
  load("../data_src/stats/sea_icemask_1982_2021avg_nc.RData")
} else {
  ## Read monthly icemask data, and must run R/01-1_OISST_monthly.Rmd first, with larger year range, e.g yrng=1982:2019
  vart <- "icemask"
  yrng <- seq(1982,2021) #2019 #old
  mmx <- 12
  gcnt <- 1 ## Cannot directly use sum/gcnt as average, because may have NA
  cntm <- matrix( rep( 0, len=1440*720), nrow = 1440)
  pret <- as.Date("1982-01-01")

  for (i in yrng) {
    for (j in 1:mmx) { ## Certainly can read a whole year at once, not necessarily read only two months
      monj <- fifelse(j<10, paste0("0",j), paste0(j))
      print(paste0("Now in Year-month: ", i," - ", monj))
      filet <- paste0("../data_src/oisst/monthly_", vart, "/", i, monj, "_", vart, ".nc")
      curt <- as.Date(paste0(i, monj, "01"), format="%Y%m%d")
    
      if (gcnt==1) { 
        stx <- stars::read_stars(filet)
        names(stx)[1] <- vart
        stx <- stx %>% select(vart)
        cntm[which(!is.na(stx[[1]]))] <- 1
        gcnt <- gcnt + 1
        next 
      }
      x <- stars::read_stars(filet) 
      names(x)[1] <- vart 
      x <- x %>% select(vart)
      cntm[which(!is.na(x[[1]]))] <- cntm[which(!is.na(x[[1]]))] + 1
      stx <- c(stx, x)
      names(stx) <- c(vart, vart)
      stx <- merge(stx) %>% st_set_dimensions(3, values = as.POSIXct(c(pret, curt)), names = "time") %>% 
          aggregate(by="2 months", FUN=sum, na.rm=TRUE)
      names(stx) <- vart
      stx <- stx %>% select(vart) %>% adrop
      if (i==yrng[length(yrng)] & j==mmx) { ## last month
        stx[[1]] <- stx[[1]]/cntm
      } else {
        pret <- curt
        gcnt <- gcnt + 1
      }
    } 
  }
}
icemask <- as.data.table(stx)
icemask[, x:=fifelse(x>180, x-360, x)]
```

``` {r merge_land_sea_icemask, warning=FALSE, message=FALSE, results='hide'}
dt <- merge(dt0, icemask, by=c("x", "y"), all=TRUE) ## this icemask is an average value through time
dt[is.na(landmask), `:=`(landmask=0, seamask=1)]
dt[seamask==1, seamask:=NA_integer_] ## reassign seamask according to average icemask
dt[, remark:=NA_character_]

## The following code ref: https://github.com/mjacox/Thermal_Displacement/blob/master/make_oisst_masks.m
## Permanent sea ice
dt[icemask>.9 & is.na(seamask), seamask := 1L]
dt[seamask==1L, remark:='Permanent sea ice']

## Seasonal sea ice
dt[icemask>0 & is.na(seamask), seamask := 2L]
dt[seamask==2L, remark:='Seasonal sea ice']

## Here we unify x: (0,360) and lon: (-180, 180)
setnames(dt, 1:2, c("lon", "lat"))
dt[,`:=`(x=fifelse(lon<0, lon+360, lon), y=lat, region=NA_character_)]

## Ice-free areas surrounded by sea ice
dt[y < -63.9 & is.na(seamask), `:=`(seamask=3, region='Antarctica')]
dt[x>12 & x<32 & y>53.5 & y<=66 & is.na(seamask), `:=`(seamask=3, region='Baltic')]
dt[x>9.9 & x<12 & y>53 & y<=60 & is.na(seamask), `:=`(seamask=3, region='Baltic')]
dt[x>36 & x<46 & y>63 & y<=67 & is.na(seamask), `:=`(seamask=3, region='White')]
dt[x>14 & x<24 & y>77 & y<=81 & is.na(seamask), `:=`(seamask=3, region='Svalbard')]
dt[x>50 & x<190 & y>60 & y<=88 & is.na(seamask), `:=`(seamask=3, region='Russia')]
dt[x>136 & x<139 & y>53 & y<=54 & is.na(seamask), `:=`(seamask=3, region='Russia')]
dt[x>158 & x<159 & y>52 & y<=54 & is.na(seamask), `:=`(seamask=3, region='Russia')]
dt[x>160 & x<163 & y>57 & y<=60 & is.na(seamask), `:=`(seamask=3, region='Russia')]
dt[x>193 & x<207.8 & y>57 & y<=68 & is.na(seamask), `:=`(seamask=3, region='Alaska')]
dt[x>207 & x<213 & y>60 & y<=62 & is.na(seamask), `:=`(seamask=3, region='Alaska')]
dt[x>228 & x<320 & y>62.5 & y<=85 & is.na(seamask), `:=`(seamask=3, region='Canada/Greenland')]
dt[x>267 & x<284 & y>51 & y<=63 & is.na(seamask), `:=`(seamask=3, region='Canada')]
dt[x>333 & x<341 & y>70 & y<=84 & is.na(seamask), `:=`(seamask=3, region='Greenland')]
dt[x>338 & x<345 & y>64.5 & y<=68 & is.na(seamask), `:=`(seamask=3, region='Iceland')]
dt[x>267 & x<285 & y>41 & y<=50 & is.na(seamask), `:=`(seamask=3, region='Great Lakes')]
dt[x>290 & x<297 & y>45 & y<=50 & is.na(seamask), `:=`(seamask=3, region='NW Atlantic')]
dt[x>302 & x<307 & y>47 & y<=54 & is.na(seamask), `:=`(seamask=3, region='NW Atlantic')]
dt[seamask==3, remark:='Ice-surrounded areas'];

## Caspian Sea
dt[x>=46 & x<=56 & y>=36 & y<=48 & is.na(seamask), `:=`(seamask=4, region='Caspian Sea')]
dt[seamask==4, remark:='Caspian Sea']

## Black Sea
dt[x>=26.8 & x<=42 & y>=40 & y<=48 & is.na(seamask), `:=`(seamask=5, region='Black Sea')]
dt[seamask==5, remark:='Black Sea']

## Mediterranean Sea
dt[x<=26.7 & y>=30 & y<=46 & is.na(seamask), `:=`(seamask=6, region='Mediterranean Sea')]
dt[x>=26 & x<=37 & y>=30.5 & y<=39.5 & is.na(seamask), `:=`(seamask=6, region='Mediterranean Sea')]
dt[x>=354 & y>=33 & y<=41 & is.na(seamask), `:=`(seamask=6, region='Mediterranean Sea')]
dt[seamask==6, remark:='Mediterranean Sea']

## Red Sea
dt[x>=32 & x<=43 & y>=12.5 & y<=30 & is.na(seamask), `:=`(seamask=7, region='Red Sea')]
dt[seamask==7, remark:='Red Sea']

## Persian Gulf
dt[x>=46 & x<56 & y>=23 & y<=31 & is.na(seamask), `:=`(seamask=8, region='Persian Gulf')]
dt[seamask==8, remark:='Persian Gulf']

## Northern Arabian Sea
dt[x>=45 & x<75 & y>=14 & y<=28 & is.na(seamask), `:=`(seamask=9, region='Northern Arabian Sea')]
dt[seamask==9, remark:='Northern Arabian Sea']

## Northern Bay of Bengal
dt[x>=77 & x<99 & y>=14 & y<=25 & is.na(seamask), `:=`(seamask=10, region='Northern Bay of Bengal')]
dt[seamask==10, remark:='Northern Bay of Bengal']

## Equatorial Indian Ocean
dt[x>=37 & x<99 & y>= -5 & y<=15 & is.na(seamask), `:=`(seamask=11, region='Equatorial Indian Ocean')]
dt[x>=99 & x<=100 & y>= -5 & y<8 & is.na(seamask), `:=`(seamask=11, region='Equatorial Indian Ocean')]
dt[x>100 & x<=101 & y>= -5 & y<6.7 & is.na(seamask), `:=`(seamask=11, region='Equatorial Indian Ocean')]
dt[x>101 & x<=104 & y>= -5 & y< -2 & is.na(seamask), `:=`(seamask=11, region='Equatorial Indian Ocean')]
dt[seamask==11, remark:='Equatorial Indian Ocean']

## South China Sea
dt[x>=98 & x<120 & y>= -5 & y<=30 & is.na(seamask), `:=`(seamask=12, region='South China Sea')]
dt[seamask==12, remark:='South China Sea']

## Northern Gulf of California
dt[x>=244.5 & x<248 & y>=29.7 & y<=32 & is.na(seamask), `:=`(seamask=13, region='Northern Gulf of California')]
dt[x>=245.5 & x<249 & y>=29.4 & y<=30 & is.na(seamask), `:=`(seamask=13, region='Northern Gulf of California')]
dt[x>=246 & x<249 & y>=28.9 & y<=30 & is.na(seamask), `:=`(seamask=13, region='Northern Gulf of California')]
dt[x>=246.5 & x<252 & y>=27 & y<=30 & is.na(seamask), `:=`(seamask=13, region='Northern Gulf of California')]
dt[x>=247.8 & x<252 & y>=26.4 & y<=30 & is.na(seamask), `:=`(seamask=13, region='Northern Gulf of California')]
dt[x>=248.2 & x<252 & y>=25 & y<=30 & is.na(seamask), `:=`(seamask=13, region='Northern Gulf of California')]
dt[seamask==13, remark:='Northern Gulf of California']

## Eastern Tropical Pacific
dt[x>=245 & x<260 & y>= 0 & y<=25 & is.na(seamask), `:=`(seamask=14, region='Eastern Tropical Pacific')]
dt[x>=255 & x<262 & y>= 0 & y<=20 & is.na(seamask), `:=`(seamask=14, region='Eastern Tropical Pacific')]
dt[x>=261 & x<270 & y>= 0 & y<=17 & is.na(seamask), `:=`(seamask=14, region='Eastern Tropical Pacific')]
dt[x>=269 & x<275.7 & y>= 0 & y<=14 & is.na(seamask), `:=`(seamask=14, region='Eastern Tropical Pacific')]
dt[x>=269 & x<282.8 & y>= 0 & y<=7.4 & is.na(seamask), `:=`(seamask=14, region='Eastern Tropical Pacific')]
dt[x>=269 & x<282 & y>= 0 & y<=8.5 & is.na(seamask), `:=`(seamask=14, region='Eastern Tropical Pacific')]
dt[x>=275 & x<277 & y>=8.4 & y<=9.5 & is.na(seamask), `:=`(seamask=14, region='Eastern Tropical Pacific')]
dt[x>=280 & x<281.6 & y>=8.4 & y<=9 & is.na(seamask), `:=`(seamask=14, region='Eastern Tropical Pacific')]
dt[x>=245 & x<295 & y>= -30 & y<= 0 & is.na(seamask), `:=`(seamask=14, region='Eastern Tropical Pacific')]
dt[seamask==14, remark:='Eastern Tropical Pacific']

## Northern Gulf of Mexico
dt[x>=260 & x<278.5 & y>=27 & y<=31 & is.na(seamask), `:=`(seamask=15, region='Northern Gulf of Mexico')]
dt[seamask==15, remark:='Northern Gulf of Mexico']

## Western Tropical Atlantic
dt[x>=260 & x<315 & y>= -10 & y<=27 & is.na(seamask), `:=`(seamask=16, region='Western Tropical Atlantic')]
dt[seamask==16, remark:='Western Tropical Atlantic']

## US East Coast
dt[x>=278.5 & x<315 & y>=27 & y<=47 & is.na(seamask), `:=`(seamask=17, region='US East Coast')]
dt[seamask==17, remark:='US East Coast']

## Treat some water bodies as land ## https://github.com/mjacox/Thermal_Displacement/blob/master/oisst_an.m
dt[x>=267 & x<285 & y>=41 & y<=50, seamask:=0] ## Great lakes #But hereafter programs use landmask (here only change seamask)
dt[x>=269.5 & x<270.5 & y>=30 & y<=31, `:=`(seamask=0, region='Lake Ponchartrain')] ## Lake Ponchartrain
print(unique(dt[seamask==0 & landmask!=1, .(remark, region)]))
#### We will use landmask -> land, new landmask is to mask land + inland lake
dt[,land:=landmask]
dt[seamask==0 & landmask!=1 & region %chin% c("Great Lakes", "Lake Ponchartrain"), landmask:=1]
```

```{r plot_land_seamask, warning=FALSE, message=FALSE, results='hide'}
colors <- rev(colorRampPalette(brewer.pal(12, "Paired"))(20))
colors[2] <- "#a2d2df" #Change color of permanent sea ice
colors[3] <- "#99FFFF" #Change color of seasonal sea ice
colors[4] <- "#dcf3ff" #Change color of ice-surrounded areas'
ggplot() +  
  geom_tile(data=dt, aes(x=lon, y=lat, fill=factor(seamask))) + 
  #scale_fill_viridis_d() +
  scale_fill_manual(values = colors) +
  #coord_equal() + 
  #xlim(c(-180, 180)) + ylim(c(-90, 90)) +
  coord_sf() + #xlim=c(-180, 180), ylim=c(-90, 90)) +
  scale_x_continuous(limits = c(-180, 180), expand = c(0, 0)) + 
  scale_y_continuous(limits = c(-90, 90), expand = c(0, 0)) + 
  xlab(NULL) + ylab(NULL) +
  ggExtra::removeGrid(x=TRUE, y=TRUE) +
  theme(
      panel.background = element_blank(), 
      panel.border = element_rect(colour = "black", fill=NA, size=0.75),
      panel.grid.major.x = element_blank(),
      panel.grid.major.y = element_blank(),
      panel.grid.minor = element_blank(),
      plot.margin = unit(c(0,0,0,0), "cm"),
      panel.spacing = unit(0,"cm"),
      strip.background = element_blank(),
      strip.text.y = element_text(angle = 0),
      axis.text.x  = element_text(family = "sans"),
      axis.title.x = element_text(family = "sans"),
      axis.title.y = element_text(family = "sans"),
      axis.text.y = element_text(family = "sans"),
      legend.key = element_blank(),
      legend.key.size = unit(0.8,"line"),
      legend.box.background = element_blank(),
      legend.title = element_text(size=10),
      legend.text =  element_text(family = "sans", size=7),
      legend.background = element_rect(fill = "transparent", colour = "transparent"),
      legend.direction="horizontal",
      legend.key.height = unit(0.2,'cm'),
      legend.position = "bottom"
  )
```

``` {r save_results_to_reload, warning=FALSE, message=FALSE, results='hide'}
## output 
if (Recalculate) {
  #dt[,`:=`(lon=NULL, lat=NULL, landmask=NULL)]
  fwrite(dt, file="../data_src/mapfiles/sea_icemask_025d.csv")
  #save(stx, file="../data_src/mapfiles/sea_icemask_1982_2019avg_nc.RData") #old
  try(system(paste0("zip -j -m -9 ", "../data_src/mapfiles/sea_icemask_025d.csv.zip ", "../data_src/mapfiles/sea_icemask_025d.csv")))
  save(stx, file="../data_src/stats/sea_icemask_1982_2021avg_nc.RData")
}
```

``` {r test_landmask, warning=FALSE, message=FALSE, results='hide'}
# Note that in an earlier version used "GLDASp4 land mask" but current version use "GLDASp5"
# The differnt version of land mask cause difference in the definition of land areas, so that sea-mask
land_mask <- fread("unzip -p ../data_src/mapfiles/sea_icemask_025d.csv.zip")[, .(lon, lat, x, y, landmask)]
setorder(land_mask, -y, x)
land_mask[,ry:=rowid(x)]
yy <- unique(land_mask[,.(y, ry)]) ## check, NOTE that in OISST.nc file, y is from Northest (89.875) to Southest (-89.875)

landm <- dcast(land_mask, x ~ ry, value.var = "landmask")
xx <- as.numeric(landm[,1]$x)
landm <- as.matrix(landm[,-1])

## Check x order is identical to OISST file
xt <- read_stars("../data_src/oisst/monthly_sst/202205_sst.nc") %>% as.data.frame()
all.equal(xx, xt$x[1:1440]) ## TRUE

initialTrial <- TRUE
if (initialTrial) { #### Just a test to check landm really masking land, can be ignored ###########
  x <- read_stars(c("../data_src/oisst/monthly_sst/202204_sst.nc", "../data_src/oisst/monthly_sst/202205_sst.nc"), name="sst")
  names(x) <- c("Apr", "May")
  ## Just check if landm mask really on land ####
  x$Apr[1, which(landm[1,]==1)]
  x$Apr[1, which(landm[41,]==1)]
  yt=data.table(y=yy$y[as.integer(names(landm[1,which(landm[1,]==1)]))]) %>% 
    .[,x:=xx[1]] %>% list(., data.table(y=yy$y[as.integer(names(landm[41,which(landm[41,]==1)]))]) %>% 
                            .[,x:=xx[41]]) %>% rbindlist(use.names = T)
  
  yt[,`:=`(lon=fifelse(x>180, x-360, x), lat=y)]
  ggplot() +  
    geom_tile(data=land_mask, aes(x=lon, y=lat, fill=factor(landmask)), alpha=0.8) + 
    geom_point(data=yt, aes(lon,lat), color="red") + #Red points should always on land (landm == 1)
    scale_fill_viridis_d() +
    coord_sf() + #xlim(c(-15,15)) + ylim(c(min(yt$y)-5, max(yt$y)+5))
    scale_x_continuous(limits = c(-15, 15), expand = c(0, 0)) + 
    scale_y_continuous(limits = c(min(yt$y)-5, max(yt$y)+5), expand = c(0, 0))
}  
```
